requirements.txt
confluent-kafka
fastapi
uvicorn
python-dotenv